perf: array buffered output
Encoding functions now emit output in new way. By default, functions
write to stdout as before, but if $out is set, they take $out to be the
name of an array variable, and append output to the array.

This has at least two benefits. Firstly, callers can encode arrays of
data and get an array back, rather than only receiving a single, joined
string. (Most of our code is fine with a single string, but this is more
flexible, and will be useful in cases where post-processing is needed).

Second, this allows callers to receive output without using sub shell to
capture output. Avoiding unnecessary forking is a big performance win.