perf: optimistically escape \n when encoding
This has a ~15% performance benefit for strings containing newlines,
like the lines of a text file. It has a slight but lesser cost for
strings without newlines. It has a utility benefit, in that callers can
specify \n as the join char, which allows them to re-split the strings
easily or output a string per-line if desired.