feat: atomic json output
The json function now collects its output in a buffer and only outputs
it all at once, or not at all if it fails. This applies both to stdout
and array-buffered output.

As well as not polluting the outputs with partial data on errors, this
is important for compostability of multiple json bash API calls, as it
means each call to json results in a single, complete JSON value being
placed in the $out array. Without this behaviour, callers would need to
carefully join data in the array to later pass it to a :raw json
argument for further use.