feat: implement json.validate
We use PCRE via grep, which should be almost as widely available as bash
itself. Invoking grep each time a validation is needed would be
expensive, so we run grep in a background process, and have a little
validation protocol that lets us send validation requests to the grep
process via pipes.