refactor: perf: support encoding from array refs
The json.encode_* functions can now encode values from an existing
array, by calling the functions with 0 args and in=arrayname. This is
significantly more efficient than passing arrays as positional arguments
when more than a few arguments are used, or when values are large.

An unfortunate result of using namerefs is that we need to make sure
local vars in the functions won't clash with var names used by callers
in out= or in= vars. So we prefix all local vars with a shortened form
of the function name, e.g. json.encode_strings strings -> _jes_strings.
Ugly, but less bad than expecting callers to carefully follow a naming
convention to avoid clashes.