fix: handle chunk count 1 in json.stream_encode_array
We were incorrectly emitting a comma before the first array element when
json_buffered_chunk_count=1 in json.stream_encode_array. This happened
because readarray invokes our callback before the just-read chunk has
been added to the buffer (with the just-read element as an argument),
so the array is empty on the first read.

We now pre-emptively add the argument element to the buffer array before
encoding. This has the dual benefit of making the chunk visible as soon
as it's read, instead of delaying it until the following chunk is ready.

That matters little for larger chunk sizes, but it's a UX improvement
when interactively watching streaming output in the terminal.