feat!: explicitly define default attributes for json()
This is mainly driven by Bash 5.2 making it painful to detect if a
potentially-unset var is an associative array. ([[ -v array[@] ]] no
longer works on 5.2). We now provide json.define_defaults to create a
named default attribute set. The name is then used with the
json_defaults option of json(). As well as removing the need to detect
associative array vars, it also means we can avoid verifying the
defaults on each call, as it's done once on creation.

BREAKING CHANGE: the json_defaults option of the json function is now
the name of a named defaults attribute set, rather than the attributes
directly.