docs: restructure grammar to make @ and ~? flags the same
We treat the @ ref marker and the ~?? flags as part of the same
syntactic component, occurring prior to the key and value. This ends up
simplifying both the grammar rules, conceptually (flags are always prior
to the thing they effect) and just the look of the args. Flags no longer
clutter the : around the type.