docs: make value flags consistent between key and val
Previously we used @= or = for val and just @ for key. In 005 we were
allowing == in the key to disambiguate the flags from the key's value.

We're now using @ and = for both key and val, and not using @= for val.
And the key requires = or @ in order to use flags, they're no longer
optional. This makes the behaviour of both sides consistent, other than
that the key can use double-char escapes for @@ == and ::.

The intent of using == as a key prefix was to distinguish keys from
values, but this actually wasn't reliable in practice. Consider the arg
 '==key:/abc=value'  The lack of a closing / makes the second half not
 syntactically valid, so it would parse as a value of '=key:/abc=value'.

This problem still exists with a single = prefix for the key, but it's
no worse. We may need to match the argument regex in multiple stages to
prevent backtracking past a matched key and start of meta section, as
bash's regexes don't allow control over backtracking.