docs: allow flags before bare keys
Keys without a = or @ prefix can now have flags in the same way as
prefixed keys. The flags are not really useful in practice for bare keys
because they're mostly expected to be constant values, but this
behaviour is consistent with the other key syntax, so should be less
surprising, and makes it easy to keep the same behaviour when adding or
removing a prefix at the command line.

(Previously, flags were considered part of bare key values, e.g. ??foo
would be parsed as "??foo":"" ).

It's also consistent with how flags at the end of keys are parsed as
part of the value. e.g. foo?? is parsed as "foo":"" with ?? as value
flags.

In both cases the flags can be made part of the key. By prefixing with
a =, and suffixing with :. e.g. =??foo??: is parsed as "??foo??":"".