feat: implement type-specific JSON validation
json.validate can now validate values as particular JSON types, like
number, string, or objects with string values, or arrays with number
values.

i.e. as well as validating that an input string is a valid JSON
document, it can validate that the document is specifically a JSON
string, or object containing only string values, etc.

We need this to implement the splat operator, and to merge object
entries when implementing the {} type marker.