refactor: encode_object is now encode_object_entries
json.encode_object_entries is a bit more general than encode_object in
that it can be used to implement stream-encoding. Basically the same
except the function is not responsible for emitting surrounding { }.

The function now supports encoding entries sourced from an array of JSON
objects.

This implements the the 001-object-values plan.