feat: implement json.stream_encode_object_entries
We support encoding object entries from file chunks, similarly to how
arrays are encoded from file chunks (where chunks are lines by default,
but can be split on other separator chars).

There are two formats which define how chunks are mapped to object
entries: `json` chunks, which are JSON objects — entries are the entries
of the object (after type). And `attrs` chunks, which are key=value
attributes using the same format as the attributes in our argument
syntax.