fix: work around bash closing FDs when starting a coproc
When `json` was called for the first time with an argument using input
from process substitution and a preceding argument started the validator
coproc, the process substitution argument would fail with file not
found.

This happens because Bash seems to have a bug that results in it closing
FDs from 63 downwards when a coproc starts, and process substitution FDs
are allocated from 63 down.

See https://savannah.gnu.org/support/index.php?110910