feat: don't split inline and var ref values by default
Previously, the default split char was \n for both file refs, var refs
and inline values. Now only files are split on \n by default, others
are split on \x00 / `''` by default, i.e. not split.

I was thinking consistency was best when previously making all array
values split on \n by default, but now that we have object as well as
array values, I see a use case for multi-line values in inline
arguments, where not splitting is the desirable and intuitive default.

For example, specifying a JSON file inline in a ... splat argument. This
is easy enough to do from a terminal, either by quoting and typing it
out, or using command substitution to cat a JSON file into an argument.
Splitting on newlines here breaks that, and requires the user to
explicitly set `/split=/`, which is rather painful as we don't have
shorthand syntax for \x00 / `''` as the split char. Making it the
default seems like the least-surprising option.