fix: harden json validator coproc interaction
The json validator grep coproc interaction was working reliably when
grep was GNU grep, but it was not gracefully handling the termination of
the coproc in all cases. When a coproc exits immediately at startup,
accessing the coproc special variables (json_validator and
json_validator_PID) is subject to race condition, because bash unsets
the variables as soon as the coproc process exits. In the previous
implementation, we read the PID with ${:?} syntax, which could fail if
the PID var was unset by bash.

We now interact with the coproc in a rather more defensive way, handling
either or both of json_validator and json_validator_PID being unset at
creation; and then also handling SIGPIPE errors when later writing to
the FD of our writable end of the coproc's stdin. This is somewhat
fiddly, as it requires trapping SIGPIPE errors, otherwise bash exits
with 141.

We now have tests covering coproc failure at startup, after startup, and
IO read & write errors. We can also restart the coproc after an IO
error to recover. This could potentially occur when a read times out
under high system load.

Finally, several local vars in json.validate were not prefixed, so they
could collide with a caller's in= reference var. They're all prefixed
now.