feat: specify grep command via JSON_BASH_GREP
The JSON_BASH_GREP envar now defines the grep command used by json.bash
when it starts a json validator co-process. The envar can be set to a :
delimited list of paths/commands to use. It defaults to "ggrep:grep", so
systems with ggrep installed with non-GNU grep will use ggrep.